pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = "Multi-agent research system with human-in-the-loop validation, built with LangGraph and Azure OpenAI"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Agustina Videla Rivero", email = "videlariveroa@gmail.com" }
]
dependencies = [
    "langgraph>=0.2.50",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
research-assistant = "research_assistant.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/research_assistant"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-m 'not integration'"
markers = [
    "integration: marks tests that require live Azure credentials",
]
filterwarnings = [
    "ignore::UserWarning:pydantic.main",
]
